+++ client/src/interfaces/Officer.ts
@@ -1,3 +1,5 @@
+import { Perm } from "./User";
+
 interface Officer {
   id: string;
   status: string;
@@ -6,6 +8,15 @@ interface Officer {
   officer_dept: string;
   callsign?: string;
   rank: string;
+}
+
+export interface OfficerLog {
+  id: string;
+  officer_id: string;
+  user_id: string;
+  started_at: string;
+  ended_at: string;
+  active: Perm;
 }
 
 export default Officer;